pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "botty-mcbotface"
version = "0.1.0"
description = "---"
requires-python = ">=3.10"
dependencies = ["gradio>=6.5.0", "openai", "python-dotenv"]

[tool.hatch.build.targets.wheel]
packages = ["main.py"]
